+++ examples/simple_training_net.c
@@ -285,7 +285,7 @@ mkldnn_status_t simple_net()
     /* AlexNet: relu
      * {BATCH, 96, 55, 55} -> {BATCH, 96, 55, 55}
      */
-    double negative_slope = 1.0;
+    float negative_slope = 1.0;
 
     int *relu_dst_sizes = conv_dst_sizes;
     float *relu_dst_buffer =
@@ -331,9 +331,9 @@ mkldnn_status_t simple_net()
      * k: 1.0
      */
     uint32_t local_size = 5;
-    double alpha = 0.0001;
-    double beta = 0.75;
-    double k = 1.0;
+    float alpha = 0.0001;
+    float beta = 0.75;
+    float k = 1.0;
 
     int32_t *lrn_dst_sizes = relu_dst_sizes;
