+++ src/cpu/jit_avx512_common_1x1_conv_kernel.hpp
@@ -37,7 +37,7 @@ struct jit_avx512_common_1x1_conv_kernel : public jit_generator {
                                 const memory_desc_wrapper &src_d,
                                 const memory_desc_wrapper &weights_d,
                                 const memory_desc_wrapper &dst_d,
-                                bool with_relu, double relu_negative_slope,
+                                bool with_relu, float relu_negative_slope,
                                 int nthreads, bool reduce_src);
 
     static status_t init_conf(jit_1x1_conv_conf_t &jcp,
